+++ src/cadenza.js
@@ -194,6 +194,7 @@ globalThis.cadenza = Object.assign(
  * @property {OperationMode} [operationMode] - The mode in which a workbook should be operated
  * @property {AbortSignal} [signal] - A signal to abort the iframe loading
  * @property {SnappingOptions} [snapping] - Passing these options enables snapping
+ * @property {boolean} [hideLegend] - Whether the legend should be hidden initially. Default is true.
  * @property {boolean} [useMapSrs] - Whether the coordinates specified in other parameters are specified in the map's SRS and the created geometry should use the map's SRS (otherwise EPSG:4326 is assumed)
  */
 
@@ -712,6 +713,7 @@ export class CadenzaClient {
       operationMode,
       signal,
       snapping,
+      hideLegend,
     } = {},
   ) {
     this.#log('CadenzaClient#createGeometry', ...arguments);
@@ -733,6 +735,7 @@ export class CadenzaClient {
     await this.#setupEditorViaPostRequests({
       additionalLayers,
       validExtentStrategy,
+      hideLegend,
     });
     await this.#setEditorStateToReady();
   }
@@ -766,6 +769,7 @@ export class CadenzaClient {
       signal,
       snapping,
       useMapSrs,
+      hideLegend,
     } = {},
   ) {
     this.#log('CadenzaClient#editGeometry', ...arguments);
@@ -791,6 +795,7 @@ export class CadenzaClient {
       additionalLayers,
       validExtentStrategy,
       geometry,
+      hideLegend,
     });
     await this.#setEditorStateToReady();
   }
@@ -827,6 +832,7 @@ export class CadenzaClient {
       signal,
       snapping,
       useMapSrs,
+      hideLegend,
     } = {},
   ) {
     this.#log('CadenzaClient#editGeometry', ...arguments);
@@ -850,6 +856,7 @@ export class CadenzaClient {
     await this.#setupEditorViaPostRequests({
       additionalLayers,
       validExtentStrategy,
+      hideLegend,
     });
     await this.#setEditorStateToReady();
   }
@@ -886,6 +893,7 @@ export class CadenzaClient {
       signal,
       snapping,
       useMapSrs,
+      hideLegend,
     } = {},
   ) {
     this.#log('CadenzaClient#editGeometry', ...arguments);
@@ -910,6 +918,7 @@ export class CadenzaClient {
     await this.#setupEditorViaPostRequests({
       additionalLayers,
       validExtentStrategy,
+      hideLegend,
     });
     await this.#createFeaturesAndEditLastCreatedFeature(features);
     await this.#setEditorStateToReady();
@@ -920,15 +929,18 @@ export class CadenzaClient {
    * @param [__namedParameters.geometry] {Geometry} - The geometry to edit
    * @param [__namedParameters.additionalLayers] {LayerDefinition[]} - Layer definitions to be imported and shown in the background, as a basis for the drawing. IMPORTANT: The Cadenza referenced with `cadenzaClient` must be configured to support the import of GeoJSON, and the (system) privileges of the corresponding user must also be set in such a way that the import of GeoJSON is possible.
    * @param [__namedParameters.validExtentStrategy] {ExtentStrategy} - Defines the initial map extent; If not given, Cadenza's default logic is used.
+   * @param [__namedParameters.hideLegend] {boolean} - Whether the legend should be hidden initially. Default is true.
    * @returns {Promise<Awaited<unknown>[]>}
    */
   #setupEditorViaPostRequests({
     geometry,
     additionalLayers,
     validExtentStrategy,
+    hideLegend,
   }) {
     const postRequests = [];
     postRequests.push(this.#setExtentStrategy(validExtentStrategy));
+    postRequests.push(this.#setLegendVisibility(!hideLegend));
     if (geometry) {
       postRequests.push(
         this.#postRequest('setGeometry', {
@@ -980,8 +992,20 @@ export class CadenzaClient {
     return this.#postRequest('requestEditFeature', objectId);
   }
 
+  /**
+   * Sets editor state to ready.
+   */
   #setEditorStateToReady() {
     return this.#postRequest('setEditorState', 'READY');
+  }
+
+  /**
+   * Toggles the visibility of the map legend.
+   *
+   * @param {boolean} value true to show the legend, false to hide it.
+   */
+  #setLegendVisibility(value) {
+    return this.#postRequest('setLegendVisibility', value);
   }
 
   /**